// packages/oidc-server/src/types.ts
const claimRequestSchema = z.record(
  z.string(),
  z.nullable(
    z.object({
      essential: z.boolean().optional(),
      value: z.string().optional(),
      values: z.array(z.string()).optional(),
    }),
  ),
);

export const claimsSchema = z.object({
  userinfo: claimRequestSchema.optional(),
  id_token: claimRequestSchema.optional(),
});

export type Claims = z.infer<typeof claimsSchema>;

export type ResponseType = (typeof RESPONSE_TYPES_SUPPORTED)[number];
export type ResponseMode = (typeof RESPONSE_MODES_SUPPORTED)[number];
export type SubjectType = (typeof SUBJECT_TYPES_SUPPORTED)[number];
export type SignatureAlg =
  (typeof ID_TOKEN_SIGNING_ALG_VALUES_SUPPORTED)[number];
export type TokenEndpointAuthMethod =
  (typeof TOKEN_ENDPOINT_AUTH_METHODS_SUPPORTED)[number];
export type GrantType = (typeof GRANT_TYPES_SUPPORTED)[number];
export type AcrValue = (typeof ACR_VALUES_SUPPORTED)[number];

export type AuthorizationRequest = {
  client_id: string;
  response_type: ResponseType;
  response_mode: ResponseMode;
  redirect_uri: string;
  scopes: string[];
  claims: Claims;
  state?: string;
  nonce?: string;
  prompt?: string;
  max_age?: number;
  ui_locales?: string;
  id_token_hint?: string;
  login_hint?: string;
  acr_values?: string;
  code_challenge?: string;
  code_challenge_method?: string;
};